additionalInputProps wasn't working properly
The additionalInputProps prop was all but useless. Any passed in prop would be promptly overwritten by the defaults. The correct approach would be merging the added props at the end of the TextInput list of props, so that the props passed in will overwrite the defaults. The only problem would be extending the default styles. To do that, I've taken any style prop that comes down in the additionalInputProps and merged it at the end of the TextInput styles. I think this will do the trick.